requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
SQLAlchemy==2.0.36
pydantic==2.9.2
pydantic-settings==2.6.1
python-multipart==0.0.9
python-dotenv==1.0.1
Pillow==11.0.0
numpy==1.24.3
onnxruntime==1.16.3
opencv-python